backend/requirements.txt
aenum==3.1.15
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==5.1.1
diskcache==5.6.3
evadb==0.3.7
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lark==1.1.7
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
pandas==2.1.1
psutil==5.9.6
psycopg2==2.9.9
py==1.11.0
pydantic==1.10.13
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
requests==2.31.0
retry==0.9.2
six==1.16.0
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
thefuzz==0.20.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.0
yarl==1.9.2
